Prepare control-plane smokes for v0.2 runtime
