+++ examples/control_plane/todo-route-continuation-lanes-smoke.py
@@ -70,8 +70,8 @@ def assert_route_continuation_lanes_filter_current_unclaimed_and_other_agents() 
                 "todo_id": "todo_blocking_route_gate",
                 "text": "[P0-review] Review route gate.",
                 "task_class": "advancement_task",
-                "claimed_by": "codex-main-control",
-                "blocks_agent": "codex-side-bypass",
+                "claimed_by": "codex-side-bypass",
+                "excluded_agents": ["codex-main-control"],
                 "route_id": "blocked-route",
                 "route_continuation_replan_required": True,
                 "route_continuation_reason": "same route has a next slice",
