+++ examples/control_plane/todo-summary-item-source-smoke.py
@@ -81,7 +81,9 @@ def assert_source_items_dedupe_and_mark_handoff_successor_ready() -> None:
                 "status": "done",
                 "done": True,
                 "task_class": "blocker",
-                "blocks_agent": "codex-product-capability",
+                "continuation_policy": "independent_handoff",
+                "excluded_agents": ["codex-product-capability"],
+                "unblocks_todo_id": SUCCESSOR_ID,
             },
             successor_item(index=3),
         ],
